requirements.txt
azure-storage-blob
langchain
openai
pymupdf
numpy
pandas
fastapi
uvicorn
pinecone-client
streamlit
requests
azure-functions
langchain-openai
langchain-community
pinecone-client
pdfplumber
pytesseract
Pillow
python-docx
uuid
faiss-cpu
numpy
tiktoken
